+++ app/src/main/java/org/mozilla/social/ui/MainActivityScreen.kt
@@ -1,8 +1,10 @@
 package org.mozilla.social.ui
 
+import android.annotation.SuppressLint
 import android.content.Context
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Icon
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -22,9 +24,10 @@ import org.mozilla.social.core.designsystem.icon.MoSoIcons
 import org.mozilla.social.core.domain.NavigationEvent
 import org.mozilla.social.core.navigation.NavigationDestination
 import org.mozilla.social.navigation.BottomBarTabs
-import org.mozilla.social.navigation.MozillaNavHost
+import org.mozilla.social.navigation.MainNavHost
 import org.mozilla.social.ui.navigationdrawer.NavigationDrawer
 
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @Composable
 fun MainActivityScreen(context: Context, navigationEvents: SharedFlow<NavigationEvent>) {
 
@@ -37,73 +40,55 @@ fun MainActivityScreen(context: Context, navigationEvents: SharedFlow<Navigation
         }.launchIn(this)
     }
 
-    val currentDestination = appState.currentNavigationDestination.collectAsState().value
-
-    MoSoScaffold(
-        modifier = Modifier,
-        snackbarHost = {
-            MoSoSnackbarHost(appState.snackbarHostState) { snackbarData, snackbarType ->
-                MoSoSnackbar(snackbarData = snackbarData, snackbarType = snackbarType)
-            }
-        },
-        floatingActionButton = {
-            when (currentDestination) {
-                NavigationDestination.Feed -> {
-                    MoSoFloatingActionButton(onClick = appState::navigateToNewPost) {
-                        Icon(
-                            MoSoIcons.plus(),
-                            stringResource(id = R.string.feed_fab_content_description)
-                        )
-                    }
-                }
-
-                else -> {}
-            }
-        },
-        bottomBar = {
-            BottomBar(
-                currentDestination = currentDestination,
-                navigateToTopLevelDestination = appState::navigateToTopLevelDestination
-            )
-        },
-        navigationDrawerState = appState.navigationDrawerState,
-        navigationDrawerContent = {
-            NavigationDrawer(
-                onSettingsClicked = appState::navigateToSettings,
-                onLoggedOut = appState::navigateToLoginScreen,
-            )
-        },
-        bottomSheetContent = { BottomSheetContent() },
-        bottomSheetVisible = appState.bottomSheetVisible.value,
-        content = {
-            MozillaNavHost(
-                appState = appState,
-                context = context,
-            )
-        }
+    MainNavHost(
+        appState = appState,
+        context = context,
     )
+
+
+//    Scaffold(
+//        modifier = Modifier,
+//        snackbarHost = {
+//            MoSoSnackbarHost(appState.snackbarHostState) { snackbarData, snackbarType ->
+//                MoSoSnackbar(snackbarData = snackbarData, snackbarType = snackbarType)
+//            }
+//        },
+//        floatingActionButton = {
+//            when (currentDestination) {
+//                NavigationDestination.Feed -> {
+//                    MoSoFloatingActionButton(onClick = appState::navigateToNewPost) {
+//                        Icon(
+//                            MoSoIcons.plus(),
+//                            stringResource(id = R.string.feed_fab_content_description)
+//                        )
+//                    }
+//                }
+//
+//                else -> {}
+//            }
+//        },
+//        bottomBar = {
+//            BottomBar(
+//                currentDestination = currentDestination,
+//                navigateToTopLevelDestination = appState::navigateToTopLevelDestination
+//            )
+//        },
+//        navigationDrawerState = appState.navigationDrawerState,
+//        navigationDrawerContent = {
+//            NavigationDrawer(
+//                onSettingsClicked = appState::navigateToSettings,
+//                onLoggedOut = appState::navigateToLoginScreen,
+//            )
+//        },
+//        bottomSheetContent = { BottomSheetContent() },
+//        bottomSheetVisible = appState.bottomSheetVisible.value,
+//        content = {
+//
+//        },
+//    )
 }
 
-@Composable
-private fun BottomBar(
-    currentDestination: NavigationDestination?,
-    navigateToTopLevelDestination: (route: NavigationDestination) -> Unit,
-) {
-    // don't show bottom bar if our current route is not one of the bottom nav options
-    if (BottomBarTabs.values()
-            .find { it.bottomBarTab.navigationDestination == currentDestination } == null
-    ) {
-        return
-    }
 
-    currentDestination?.let {
-        MoSoBottomNavigationBar(
-            currentDestination = currentDestination,
-            bottomBarTabs = BottomBarTabs.values().map { it.bottomBarTab },
-            navigateTo = navigateToTopLevelDestination
-        )
-    }
-}
 
 @Composable
 private fun BottomSheetContent() {
